feat: move lambda entry functions out of bin
