+++ src/bin/get-product.rs
@@ -1,12 +1,9 @@
 use lambda_http::{
-    ext::RequestExt,
     handler,
     lambda_runtime::{self, Context},
-    IntoResponse, Request, Response,
+    Request,
 };
-use products::{utils::*, Service};
-use serde_json::json;
-use tracing::{error, info, instrument, warn};
+use products::{utils::*, entrypoints::lambda_apigateway::get_product};
 
 type E = Box<dyn std::error::Error + Send + Sync + 'static>;
 
@@ -38,58 +35,4 @@ async fn main() -> Result<(), E> {
     }))
     .await?;
     Ok(())
-}
-
-/// Get a product
-#[instrument(skip(service))]
-async fn get_product(
-    service: &Service,
-    event: Request,
-    _: Context,
-) -> Result<impl IntoResponse, E> {
-    // Retrieve product ID from event.
-    //
-    // If the event doesn't contain a product ID, we return a 400 Bad Request.
-    let path_parameters = event.path_parameters();
-    let id = match path_parameters.get("id") {
-        Some(id) => id,
-        None => {
-            warn!("Missing 'id' parameter in path");
-            return Ok(response(
-                400,
-                json!({ "message": "Missing 'id' parameter in path" }).to_string(),
-            ));
-        }
-    };
-
-    // Retrieve product
-    info!("Fetching product {}", id);
-    let product = service.get_product(id).await;
-
-    // Return response
-    //
-    // Since the service returns an `Option` within a `Result`, there are three
-    // potential scenarios: the product exists, it doesn't exist, or there was
-    // an error.
-    Ok(match product {
-        // Product exists
-        Ok(Some(product)) => Response::builder()
-            .status(200)
-            .header("content-type", "application/json")
-            .body(json!(product).to_string())
-            .unwrap(),
-        // Product doesn't exist
-        Ok(None) => {
-            warn!("Product not found: {}", id);
-            response(404, json!({"message": "Product not found"}).to_string())
-        }
-        // Error
-        Err(err) => {
-            error!("Error fetching product: {}", err);
-            response(
-                500,
-                json!({"message": "Error fetching product"}).to_string(),
-            )
-        }
-    })
 }
